move lnaddr.py to lib
