+++ gui/qt/request_list.py
@@ -126,7 +126,7 @@ class RequestList(MyTreeWidget):
             self.addTopLevelItem(item)
         # lightning
         for payreq_key, r in self.wallet.lnworker.invoices.items():
-            from electrum.lightning_payencode.lnaddr import lndecode
+            from electrum.lnaddr import lndecode
             import electrum.constants as constants
             lnaddr = lndecode(r, expected_hrp=constants.net.SEGWIT_HRP)
             amount_sat = lnaddr.amount*COIN if lnaddr.amount else None
